Change #1
I made several changes to the project.  Some of the changes are that I
created the main game screen and the leader board.  I also renamed some
of the files and deleted any unneeded information from the files.  Also,
I created a css file for the main game screen.  It is under
bin/tetris/view.